+++ pkg/namespaced/controller/projects/projects/project.go
@@ -200,7 +200,7 @@ func (e *external) Create(ctx context.Context, mg resource.Managed) (managed.Ext
 		return managed.ExternalCreation{}, errors.Wrap(err, errCreateFailed)
 	}
 
-	meta.SetExternalName(cr, strconv.FormatInt(int64(prj.ID), 10))
+	meta.SetExternalName(cr, strconv.FormatInt(prj.ID, 10))
 	return managed.ExternalCreation{}, errors.Wrap(err, errKubeUpdateFailed)
 }
